fix(google_ads): id nao entra cru em GAQL, e conversao fracionada nao vira zero
Seis sitios interpolavam ad_group_id/campaign_id/ad_id como texto livre
direto na clausula IN do GAQL (quatro em _common.py, um em
audit_quality_score.py entre aspas simples, um em audit_zombie_keywords.py).
Todos passam a validar via `str(int(x))`, o idioma ja usado nos helpers de
conversion_action.id e ad_group_criterion.criterion_id no mesmo arquivo —
id invalido agora estoura ValueError antes de a query ser montada, em vez
de virar clausula GAQL arbitraria.

audit_zombie_keywords truncava metrics.conversions com int() em dois
pontos (parser da linha GAQL + boundary dict->dataclass), e o dataclass
KeywordRow/ZombieKeyword.conversions seguia tipado int. int(0.9) == 0, e
numa tool que define zumbi como "zero atividade" isso inventava zumbi:
a keyword converteu (atribuicao fracionada e o caso normal do Google) e o
relatorio dizia que nao. Os tres pontos passam a float.

+++ src/google_ads/flag_zombie_keywords.py
@@ -24,7 +24,7 @@ class KeywordRow:
     impressions: int
     clicks: int
     cost_brl: float
-    conversions: int
+    conversions: float
     status: str  # "ENABLED" expected (server-side filter)
 
 
@@ -42,7 +42,7 @@ class ZombieKeyword:
     impressions: int
     clicks: int
     cost_brl: float
-    conversions: int
+    conversions: float
     status: str
